+++ instruction_graph/if.c
@@ -93,13 +93,13 @@ void eval_ENDIF(i_graph **i_current) {
         i_level i_if = i_level_pop_branch_eval(true);
 
         i_level_set_reserved_jump(i_if.reserved_jmp_idx,
-                                  ((int64_t)i_else.i_num - (int64_t)i_if.i_num) + 1);
+                                  (i_else.i_num - i_if.i_num) + 1);
 
         // ENDING ELSE
         oper_regs_store_drop();
         reg_set *r_set = get_reg_set();
         i_level_set_reserved_jump(i_else.reserved_jmp_idx,
-                                  ((int64_t)asm_get_i_num() - (int64_t)i_else.i_num) + 1);
+                                  (asm_get_i_num() - i_else.i_num) + 1);
 
         stack_ptr_clear();
     } else if (i_level_pop_branch_eval(false).type == i_IF) {
@@ -114,10 +114,10 @@ void eval_ENDIF(i_graph **i_current) {
         i_level i_endif = i_level_pop_branch_eval(true);
 
         i_level_set_reserved_jump(i_if.reserved_jmp_idx,
-                                  ((int64_t)i_endif.i_num - (int64_t)i_if.i_num) + 1);
+                                  (i_endif.i_num - i_if.i_num) + 1);
         oper_regs_store_drop();
         i_level_set_reserved_jump(i_endif.reserved_jmp_idx,
-                                  ((int64_t)asm_get_i_num() - (int64_t)i_endif.i_num) + 1);
+                                  (asm_get_i_num() - i_endif.i_num) + 1);
 
         stack_ptr_clear();
     } else {
